직원목록 하다말앗

+++ src/main/java/com/codenal/admin/repository/AdminRepository.java
@@ -5,7 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.codenal.employee.domain.Employee;
 // 데이터베이스에 접근하는 역할
 // CRUD
-public interface AdminRepository extends JpaRepository<Employee, Integer>{
+public interface AdminRepository extends JpaRepository<Employee, Long>{
+	
+	
 
 	
 }
